requirements_py38.txt
#requests==2.32.4
#beautifulsoup4==4.6.0
#pandas==2.0.3
#mysql-connector-python==9.0.0
#python-dotenv==9.0.0
#numpy==1.24.4
#scikit-learn==1.3.2
#matplotlib==3.7.5
#seaborn==0.13.2
#konlpy==0.5.2
#jpype1==1.2.0 # konlpy의 의존성
#openpyxl==3.1.5 # pandas가 .xlsx 파일을 읽고 쓰기 위해 필요
#keybert==0.9.0 # KeyBERT 사용을 위해 필요
#sentence-transformers==3.2.1 # KeyBERT의 의존성


requests==2.32.4
beautifulsoup4==4.6.0
pandas==2.0.3
mysql-connector-python==9.0.0
python-dotenv==1.0.1
numpy==1.24.4
scikit-learn==1.3.2
matplotlib==3.7.5
seaborn==0.13.2
konlpy==0.5.2
JPype1==1.2.0
openpyxl==3.1.5
keybert==0.9.0
sentence-transformers==3.2.1
torch==2.4.1
transformers==4.46.3
scipy==1.10.1
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
urllib3==2.2.3
cffi==1.17.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
joblib==1.4.2
threadpoolctl==3.5.0
contourpy==1.1.1
cycler==0.12.1
kiwisolver==1.4.7
pillow==10.4.0
huggingface-hub==0.33.2
safetensors==0.5.3
tokenizers==0.20.3
regex==2024.11.6
lxml==6.0.0
six==1.17.0
packaging==25.0
tweepy==3.10.0